// src/spec/threatgile.generated.ts
export interface Threagile {
// ...
  shared_runtimes: {
    [k: string]: {
      /**
       * ID
       */
      id: string;
      /**
       * Description
       */
      description: string | null;
      /**
       * Tags
       */
      tags?: string[] | null;
      /**
       * Technical assets running
       */
      technical_assets_running: string[] | null;
      [k: string]: unknown;
    };
  };
// ...
}